+++ test/ibmq/test_ibmq_client_v2.py
@@ -70,10 +70,39 @@ class TestIBMQClient(QiskitTestCase):
 
         # Run the job through the IBMQClient directly.
         api = backend._api
-        job = api.run_job(qobj.as_dict(), backend_name)
+        job = api.submit_job(qobj.as_dict(), backend_name)
 
         self.assertIn('status', job)
         self.assertIsNotNone(job['status'])
+
+    @requires_qe_access
+    @requires_new_api_auth
+    def test_run_job_object_storage(self, qe_token, qe_url):
+        """Test running a job against a simulator using object storage."""
+        IBMQ.enable_account(qe_token, qe_url)
+
+        # Create a Qobj.
+        backend_name = 'ibmq_qasm_simulator'
+        backend = IBMQ.get_backend(backend_name)
+        circuit = transpile(self.qc1, backend, seed_transpiler=self.seed)
+        qobj = assemble(circuit, backend, shots=1)
+        print('complied')
+
+        # Run the job through the IBMQClient directly using object storage.
+        api = backend._api
+        job = api.job_submit_object_storage(backend_name, qobj.as_dict())
+        job_id = job['id']
+        self.assertEqual(job['kind'], 'q-object-external-storage')
+
+        # Wait for completion.
+        api.job_final_status_websocket(job_id)
+
+        # Fetch results and qobj via object storage.
+        result = api.job_result_object_storage(job_id)
+        qobj_downloaded = api.job_download_qobj_object_storage(job_id)
+
+        self.assertEqual(qobj_downloaded, qobj.as_dict())
+        self.assertEqual(result['status'], 'COMPLETED')
 
     @requires_qe_access
     @requires_new_api_auth
@@ -133,7 +162,7 @@ class TestIBMQClient(QiskitTestCase):
         qobj = assemble(circuit, backend, shots=1)
 
         api = backend._api
-        job = api.run_job(qobj.as_dict(), backend_name)
+        job = api.submit_job(qobj.as_dict(), backend_name)
         job_id = job['id']
 
         # Get the job, excluding a parameter.